Migrations

+++ wipe_db.py
@@ -1,10 +1,18 @@
 # wipe_db.py
+import os
+
+from sqlalchemy import create_engine
+
 from db.models import Base
-from bot.db import sync_engine
 
 
 def wipe() -> None:
-    with sync_engine.begin() as conn:
+    database_url = os.getenv("DATABASE_URL")
+    if not database_url:
+        raise RuntimeError("DATABASE_URL not found")
+
+    engine = create_engine(database_url)
+    with engine.begin() as conn:
         Base.metadata.drop_all(bind=conn)
         Base.metadata.create_all(bind=conn)
